Split roles of [. : /]
Rather than a single PATH! type, Ren-C now has three interstitial
delimiter types:

   tu.p.le
   ch:a:in
   pa/th

The heavier the delimiter, the "outermost" it is in the structure:

   a.b:c/d:e.f

That is a 2-element PATH! [a.b:c d:e.f], the first of which is a
CHAIN! [a.b c] and the second of which is also a CHAIN! [d e:f]

While trailing colons as decoration are still a "SET" intent, the
leading colon is now the intent used for "refinements".

   >> append:dup [a b c] [d e]
   == [a b c [d e] [d e]]

Slashes serve the role of indicating function application, while
tuple is for "inert" picking.

Also moves away from the need for APPEND:ONLY or INSERT:ONLY, thanks
to the advent of SPLICE!:

  https://rebol.metaeducation.com/t/the-long-awaited-death-of-only/1607