FAIL is not divergent, use PANIC for divergence
The word FAIL has been repurposed to generate definitional errors,
usually as `return fail ...`

This was originally called RAISE, but FAIL is believed to be an
improvement.

If your code notices an alarming condition that's meant to stop the
program, use PANIC.

Note also that ERROR! is now the name of the unstable antiform (as
opposed to RAISED!).  The stable form is called WARNING!

Discussion:

  https://rebol.metaeducation.com/t/panic-vs-fail-raised/2424