BINARY! renamed to BLOB!
There are multiple reasons for this naming change.  One is that the
overloaded term "binary" often implies base-2 notation to people, or
alternately an executable or object file on disk.

Also, internally to the implementation, it was necessary to have a
distinct name for the underlying storage of bytes from the cell that
is exposed to the user.

Since "BLOB" is a shorter name than "BINARY" and is narrower in meaning
to arbitrary byte data in programming, it was switched.  Internally to
the C code, `Binary` is the name used to describe the memory for the
bytes (no series index, no cell header, just the data).  This helps
conveniently differentiate functions like Binary_At() from Blob_At(),
etc.

Discussion:

  https://rebol.metaeducation.com/t/binary-vs-bytes-vs-blob/1477/