EVAL instead of DO for code with no header
The roles of EVAL and DO are split now.

EVAL is specific to Rebol code, and expects the code to already be
bound appropriately.  It does not accept string input.  It can go
step-by-step in evaluations with EVAL:STEP

DO offers no stepwise operation, and expects to be given enough
information to assemble any needed binding for any stringlike material
you are trying to execute.  If given a string, it has to have a header.

Future directions are to make DO a "polyglot", and able to execute
CSS or JS... eliminating the need for CSS-DO and JS-DO.

Discussion: