"Scope detection" instead of reb.Arg/reb.ArgR
Modern binding is able to project an environment onto the execution
of code that would otherwise be bound, and this is now integral to
the librebol API.  So you no longer need reb.Arg() or reb.ArgR().

In JavaScript in particular, there is a *per-JS-NATIVE* customization
of what the `reb` object is that you're calling APIs out of.  Each
function has its own `reb`, and that knows how to look up parameters
that should be visible just in that function.

This is important, for instance if you write:

   js-do --[
       function some_helper(p) {
          reb.Elide("print [-[print works!]-", rebI(p), "]")
       }
   ]--

   demo: js-native [print [integer!]] --[
       let p = reb.UnboxInteger("print")
       some_helper(p)
   ]--

The `reb` you use in DEMO knows to consider the arguments and use the
name PRINT from there, but when you call the subroutine that's not
inside a JS-NATIVE, so it uses the `reb` outside the function (which
may in turn, be constrained to knowing the definitions in a specific
module).

When you override variables in this way it is called "variable
shadowing", and it's the cornerstone of how Ren-C is able to mirror
some of the hierarchical environment we can accomplish within the
interpreter when interacting with APIs.

Discussion:

  https://rebol.metaeducation.com/t/scope-detection-in-js-and-c/2157