TRAP vs SYS.UTIL.RECOVER
TRAP is now a very specialized function for propagating ERROR! antiforms
that come back from a function call to current definitional RETURN:

  https://rebol.metaeducation.com/t/virtual-error-propagation/2389

It's not clear what was intended by having TRAP on a READ SRC vs.
allowing it to PANIC normally.  But as a talking point, this keeps the
TRAP in the EDIT function such that if a READ happens and it returns
a definitional ERROR!, that will be propagated so that the EDIT will
return that erro.  This means you can say `try edit ...` and it would
be able to defuse the ERROR! and become null on a failed read.

The other use of TRAP from the pre-antiform-ERROR!-era was to catch
arbitrary evaluation errors, regardless of stack depth, to report
them in the evaluation.  It's one of the few places that intercepting
a PANIC is valid.

Remarks on why RESCUE is hidden in SYS.UTIL, and why the likely only
reasonable response to a PANIC is to report it:

  https://rebol.metaeducation.com/t/discouraging-interception/1871

There is also some flux surrounding the TO operation being willing
to canonize conversions of UTF-8 BLOB! to text and back.  This is
a work in progress, still being decided.  In the meantime, ENCODE
and DECODE of 'UTF-8 works... and you can AS alias a BLOB! as a
TEXT! in order to take over the internal memory (not appropriate if
you don't want modifications to the text to change the binary, but
if you did a READ of that data there are no other references, so you
might as well use that memory to back your string).