Resolve script load race conditions: JS-AWAITER
We don't want the ADD-EDITOR function to return to its Rebol caller
until it has loaded the ACE editor successfully.

To accomplish this, instead of using a JS-NATIVE we use a JS-AWAITER.
This returns a Promise(), which means that the JavaScript extension
will not continue until either the resolve() or reject() functions of
that Promise are called.

The script.onload() and script.onerror() functions are then hooked for
the ACE editor script, to call resolve() and reject(), to unblock the
execution when ready.

(Also changes the ACE editor to use cloudflare instead of pagecdn.io,
as pagecdn.io now appears to be a spam/malware site.)

Same applies to JS frame and panel.