Switch DB/GET to return NULL vs. "trash"
At the time that the rebol-web-scripts were initially created, the
meaning of reb.Void() was probably somewhat similar to what today would
be the "tripwire TRASH! antiform".

Today, the correct answer for something like a DB/GET which can't
find what you're looking for would either be NULL or an ERROR!
antiform (which you can TRY to turn into a NULL, or use other error
handling with things like EXCEPT or TRAP...)

For convenience in the moment, this just returns null.  But if you
are going to do tests on a null (such as TEXT?) then you have to OPT
that NULL, to produce a VOID, so as to "opt out" of the test.  Because
tests like TEXT? want to be more careful about letting you think you've
tested a valid value these days.

Discussions:

  https://rebol.metaeducation.com/t/things-like-text-fail-on-null/2452

  https://rebol.metaeducation.com/t/range-check-error/1825