FOR-EACH, FOR-NEXT, etc. take arguments quoted
There's long been a question over whether the argument to things like
FOR-EACH should be quoted...or if they should be taken as "literal
arguments" (escapable via GROUP!, if you wanted an evaluation to
produce the name of the loop variable)

Ren-C makes the odd choice that it's a literal argument, but also
expected to carry a quote mark.  The reason it is literal is because
we want to see other decorators (like $x or ^x) without having to
quote them.

Discussion:

  https://rebol.metaeducation.com/t/literal-proxies-for-dialects/2489