EXPORT intended module definitions
The current behavior of DO is to isolate all of a script's code and
just return you the last evaluation of that script (or whatever it
passed to QUIT).

This means that if you want to expose functions from inside of a
script to the outside, you have to make it a module, and say which
declarations are exported (either in the header, as an EXPORT on a
SET-WORD!, or using the export function on a block of definitions).

So if you're going to add commands to the environment to be called
via the ReplPad, right now that's generally done with IMPORT of
the file defined as `type: module`

The ergonomics of this, or alternative ways to integrate applications
into the ReplPad environment, are up for discussion.