true/false/on/off/yes/no WORD!s, not variables
The new concept of "flexible logic" goes to the idea that the only
falsey value is the null antiform--which cannot be put in a block.
There is a complement "okay" antiform, but all non-antiforms are
also truthy.

Consequently, there's nothing you can put in a BLOCK! that is "falsey"
in and of itself.  This is a system is called "flexible logic".

Discussion:

  https://rebol.metaeducation.com/t/flexible-logic-terminology/2252