Simplify evaluation in %editor.r for now (broken)
To evaluate arbitrary Rebol code, that code may interact with the
browser, and requires use of Promises/JS-AWAITER mechanics that are
somewhat complex.  The ReplPad does it, but it's not as easy as it
should be for other code to do it.

This switches to a simple evaluation that can do things which do not
need to run JS-AWAITER functions.  But it's broken at the moment, bug
is under investigation.  Stay tuned...